# docs/swagger/swagger.yaml
basePath: /gateway
definitions:
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.CreatedHotelResponse:
    properties:
      address:
        type: string
      description:
        type: string
      id:
        type: string
      message:
        type: string
      name:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.CreatedRoomResponse:
    properties:
      id:
        type: string
      message:
        type: string
      number:
        type: string
      room_type_id:
        type: string
      status:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.CreatedRoomTypeResponse:
    properties:
      amenities:
        type: string
      base_price:
        type: number
      capacity:
        type: integer
      hotel_id:
        type: string
      id:
        type: string
      message:
        type: string
      name:
        type: string
    type: object
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.HotelRequest:
    properties:
      address:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.HotelResponse:
    properties:
      address:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      room_types:
        items:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomTypeSummary'
        type: array
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.HotelUpdateRequest:
    properties:
      address:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomRequest:
    properties:
      number:
        type: string
      room_type_id:
        type: string
      status:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomResponse:
    properties:
      id:
        type: string
      number:
        type: string
      room_type_id:
        type: string
      status:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomTypeRequest:
    properties:
      amenities:
        type: string
      base_price:
        type: number
      capacity:
        type: integer
      hotel_id:
        type: string
      name:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomTypeResponse:
    properties:
      amenities:
        type: string
      base_price:
        type: number
      capacity:
        type: integer
      hotel_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomTypeSummary:
    properties:
      capacity:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomUpdateRequest:
    properties:
      number:
        type: string
      status:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.SuccessResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8088
info:
  contact: {}
  description: Aggregated API surface for hotel booking platform.
  title: Hotel Booking Microservices API
  version: "1.0"
paths:
...
  /hotels:
    get:
      parameters:
      - description: pagination limit (default 50)
        in: query
        name: limit
        type: integer
      - description: pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.HotelResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      summary: List hotels
      tags:
      - Hotels
    post:
      consumes:
      - application/json
      parameters:
      - description: Hotel payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.HotelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.CreatedHotelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create hotel
      tags:
      - Hotels
  /hotels/{id}:
    delete:
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete hotel
      tags:
      - Hotels
    get:
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.HotelResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      summary: Get hotel detail
      tags:
      - Hotels
    put:
      consumes:
      - application/json
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      - description: Hotel update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.HotelUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.HotelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update hotel
      tags:
      - Hotels
...
  /room-types:
    get:
      parameters:
      - description: pagination limit (default 50)
        in: query
        name: limit
        type: integer
      - description: pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomTypeResponse'
            type: array
      summary: List room types
      tags:
      - Hotels
    post:
      consumes:
      - application/json
      parameters:
      - description: Room type payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.CreatedRoomTypeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create room type
      tags:
      - Hotels
  /rooms:
    get:
      parameters:
      - description: pagination limit (default 50)
        in: query
        name: limit
        type: integer
      - description: pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomResponse'
            type: array
      summary: List rooms
      tags:
      - Hotels
    post:
      consumes:
      - application/json
      parameters:
      - description: Room payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.CreatedRoomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create room
      tags:
      - Hotels
  /rooms/{id}:
    delete:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete room
      tags:
      - Hotels
    get:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      summary: Get room by ID
      tags:
      - Hotels
    put:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      - description: Room update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update room
      tags:
      - Hotels
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"